Add @Step description and @DisplayName for tests. Code improvement

+++ src/test/java/GetUserOrdersTest.java
@@ -1,25 +1,48 @@
+import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
-import org.junit.Assert;
+import order.OrderHelper;
+import org.junit.After;
 import org.junit.Test;
+import user.Random;
+import user.User;
+import user.UserHelper;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
 
 public class GetUserOrdersTest {
-    String accessToken;
-    OrderHelper orderHelper = new OrderHelper();
+    private String accessToken;
+    private final OrderHelper orderHelper = new OrderHelper();
 
     @Test
+    @DisplayName("Получение заказов пользователя с авторизацией")
     public void getOrdersAuthorized() {
         Random random = new Random();
         User user = random.getRandomUser();
         Response response = UserHelper.postCreateUser(user);
         accessToken = response.then().extract().path("accessToken").toString().substring(6).trim();
-        boolean success = orderHelper.getUsersOrdersToken(accessToken);
-        Assert.assertTrue(success);
-        UserHelper.deleteUser(accessToken);
+        response = orderHelper.getUsersOrdersToken(accessToken);
+        response.then()
+                .assertThat()
+                .statusCode(200)
+                .and()
+                .body("success", is(true));
     }
 
     @Test
+    @DisplayName("Получение заказов пользователя без авторизации")
     public void getOrdersNotAuthorized() {
-        String expected = orderHelper.getUsersOrders();
-        Assert.assertEquals("You should be authorised", expected);
+        Response response = orderHelper.getUsersOrders();
+        response.then()
+                .assertThat()
+                .statusCode(401)
+                .and()
+                .body("message", equalTo("You should be authorised"));
+    }
+
+    @After
+    public void cleanUp() {
+        if (accessToken != null) {
+            UserHelper.deleteUser(accessToken);
+        }
     }
 }
